+++ NEWS.md
@@ -5,6 +5,8 @@ fit_result objects as a data frame.
 contained within a list structure.
 * Basis set and noise region are now checked for validity in ABfit.
 * Added spant_abfit_benchmark function.
+* Tentative functions for performing "standard" 1H brain analyses: 
+svs_1h_brain_analysis and svs_1h_brain_batch_analysis.
 
 # spant 1.10.0
 * Fix for NIfTI MRS reader/writer.